% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_functions_20210517.R
\name{defineRegion_UTR}
\alias{defineRegion_UTR}
\title{Generate units based on UTR regions according to given downstream and upstream base pair numbers}
\usage{
defineRegion_UTR(
  up5UTR_bp = 1000,
  down3UTR_bp = 1000,
  gtf_border,
  regUnit_filename
)
}
\arguments{
\item{down3UTR_bp}{number of base pairs downstream of 3' UTR, default to 1000}

\item{gtf_border}{dataframe of genes' border info}

\item{regUnit_filename}{output filename of userdefined regions}

\item{up5TUR_bp}{number of base pairs upstream of 5' UTR, default to 1000}
}
\description{
Generate units based on UTR regions according to given downstream and upstream base pair numbers
}
\keyword{GTF}
\keyword{defined}
\keyword{file}
\keyword{from}
\keyword{get}
\keyword{regions}
\keyword{regulatory}
\keyword{user}
