% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_functions_20210517.R
\name{get_geneborder}
\alias{get_geneborder}
\title{Generate gene borders according to the gtf file given a list of genes of interest}
\usage{
get_geneborder(gtf_df, geneList, geneBorder_filename)
}
\arguments{
\item{gtf_df}{parsed gtf datafrome, can be loaded from the package}

\item{geneList}{a list of gene symbols of interest}

\item{geneBorder_filename}{output filename of borders of genes of interest}
}
\description{
Generate gene borders according to the gtf file given a list of genes of interest
}
\keyword{a}
\keyword{border}
\keyword{coordinates}
\keyword{genes}
\keyword{get}
\keyword{ghe}
\keyword{list}
\keyword{of}
