% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_functions_20210517.R
\name{get_protGeno}
\alias{get_protGeno}
\title{Generate prot units with genome coordinates}
\usage{
get_protGeno(protUnit_filename, mappedProtUnit_filename)
}
\arguments{
\item{protUnit_filename}{file of protein units of interest}

\item{mappedProtUnit_filename}{output filename of protein units with genome coordinates added}
}
\description{
Generate prot units with genome coordinates
}
\keyword{a}
\keyword{border}
\keyword{coordinates}
\keyword{genes}
\keyword{get}
\keyword{list}
\keyword{of}
\keyword{the}
