% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_inputWorkflow_function_20210724.R
\name{gpd_workflow}
\alias{gpd_workflow}
\title{Compare between 2 conditions (paired) in 1D}
\usage{
gpd_workflow(
  vcf_folderPath,
  mapping_vcf_to,
  mapTo_fileName,
  gtf_fileName,
  reg_fileName,
  ud_fileName,
  output_folderPath,
  output_tag
)
}
\arguments{
\item{vcf_folderPath}{names of the columns in the first condition}

\item{mapping_vcf_to}{names of the columns in the second condition}

\item{mapTo_fileName}{dataframe for the data, rows are genes, columns are patients in 2 conditions}

\item{gtf_fileName}{genes with values in at least the number of samples are included in comparison}

\item{reg_fileName}{seeds for the permutation}

\item{ud_fileName}{number of permutations}

\item{output_tag}{a label for the anlaysis}

\item{output_folderPaht}{the directory output files will be deposited in}
}
\description{
Compare between 2 conditions (paired) in 1D
}
\keyword{mapping}
