% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_inputWorkflow_function_20210724.R
\name{gpd_workflow}
\alias{gpd_workflow}
\title{mapping vcf to various units}
\usage{
gpd_workflow(
  vcf_folderPath,
  grab_start_string = NULL,
  grab_sep = NULL,
  grab_number = NULL,
  mapping_vcf_to,
  mapTo_fileName,
  gtf_df,
  reg_fileName,
  ud_fileName,
  output_folderPath,
  output_tag
)
}
\arguments{
\item{vcf_folderPath}{the path to the folder of vcf files}

\item{grab_start_string}{a string indicating the sample's name}

\item{mapping_vcf_to}{mapping mode, one of "GTF","regulatory","protUnits","userDefine"}

\item{mapTo_fileName}{file of protein units to be mapped with genome coordinates}

\item{gtf_df}{gtf dataframe to be mapped, can be loaded with the package}

\item{reg_fileName}{file of regulatory units to be mapped}

\item{ud_fileName}{file of user-defined regions}

\item{output_tag}{a tag for output files}

\item{output_folderPaht}{the directory holding output files}
}
\description{
mapping vcf to various units
}
\keyword{mapping}
\keyword{to}
\keyword{units}
\keyword{various}
\keyword{vcf}
